search for lpeg.so on MacOS
Currently Neovim only searches for lpeg.dylib, liblpeg.dylib, etc. on MacOS, but a normal installation of lpeg will produce lpeg.so instead.  There are explicit workarounds for this by modifying lpeg in Homebrew, and will shortly be another in spack.  Fixing it upstream as well maybe we can drop the workarounds at some point.